refactor: Rename claim to lodgement
Changes made are to avoid confusion, since so far it has meant "claiming" on the DataHighway that you've obtained a transaction hash as a result of locking or signaling some tokens and you want to record it on the DataHighway so your claims eligibility may be checked.

Instead we will refer to this action as a "lodgement" of a commitment (a commitment that you've locked or signaled tokens for a period of time).

The reason why is because it should more appropriately be used to mean you're "claiming" your tokens back at the end of the lock period, as a result of a return request (a request for your locked or signaled tokens to be returned).